remove the frame_controller class
All the frame controller does is calls one function of a frame_generator
and spits out the result on a cadance. It makes a lot more sense for this
functionality to be part of the frame_generator base class. This commit
moves all functionality of the frame_controller class into the frame_generator
class.