Don't compile C code as C++
The compiler yells at us, as it should. This was originally a hack
because we didn't want to figgure out linking C++ with C at 4am, but
this commit introduces a fix. Add an 'extern "C" {' wrapper to the
declarations in piHelpers.h when it's beint interpreted by a C++
compiler, and fix the Makefile to actually compile it a C and not C++.

The code compiles and links with these changes, and the compiler no longer
yells at us for trying to compiles C with a C++ compiler. Yay.