Use the right size for c_sample in scrolling_fft_generator::make_next_frame
We want the size of c_sample to be a power of 2 so that fft is happy. The
previous code failed to do this and caused memory corruption and infinite
loops, etc. This commit fixes this. Tested by running
./visualize space_odity.wav and ensuring that the program finished without
segfaulting in about the same amount of time that the .wav file takes to play.