Replace trivial_frame_generator with something more useful.
Replace trivial frame generator with lambda_generator. This class
takes a lambda as a constructor argument. That lambda is then
called when 'make_next_frame' is called. This allows for easy
prototyping without having to deal with annoyance of writing a whole
class.

Similarly, change visualize.cpp to use this new class instead of
trivial_frame_generator.